+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 - BitmapText: fix improper `fillStyle` initialization in BitmapText (thanks @wpernath)
 - Canvas: fix the "transparent" canvas mode with the canvas renderer (thanks @wpernath)
+- Sprite: properly apply a tint when passed through the settings to the constructor
 - WebGL: fix a regression with global opacity not being properly cascaded to texture in WebGL mode (thanks @wpernath)
 
 ## [13.1.1] (melonJS 2) - _2022-08-10_